requirements.txt
matplotlib==3.5.1
numpy==1.21.5
openpyxl==3.0.9
pandas==1.4.2
Pillow==9.1.1
plotly==5.6.0
streamlit==1.10.0
st_aggrid